requirements.txt
asgiref==3.5.0
autopep8==1.6.0
distlib==0.3.4
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
filelock==3.6.0
pbr==5.8.1
Pillow==9.1.0
platformdirs==2.5.1
pycodestyle==2.8.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
stevedore==3.5.0
toml==0.10.2
tzdata==2021.5
virtualenv==20.13.4
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
